+++ yuniql-tests/platform-tests/MigrationServiceFactory.cs
@@ -68,7 +68,7 @@ namespace Yuniql.PlatformTests
 
             var configurationService = new ConfigurationDataService(dataService, _traceService, tokenReplacementService);
 
-            var migrationService = new MigrationService(
+            var migrationService = new TransactionalMigrationService(
                 localVersionService,
                 dataService,
                 bulkImportService,
